Expire and makehistory will not check if the message-id already exists in the database if they get an error with dbzstore.  If the message-id is duplicate then just print it as an error and skip to the next line. Unfortunately older versio had bugs in art.c that would cause TRASH to be REMEMBERED muliple times.
Obtained from:

+++ expire/makehistory.c
@@ -165,11 +165,15 @@ Rebuild(size, IgnoreOld, Overwrite)
 	key.dptr = p;
 	key.dsize = save - p + 1;
 	if (!dbzstore(key, value)) {
-	    (void)fprintf(stderr, "Can't store \"%s\", %s\n",
-		    p, strerror(errno));
-	    if (temp[0])
-		(void)unlink(temp);
-	    exit(1);
+	    if (dbzexists(key)) {
+	        fprintf(stderr, "Duplicate message-id \"%s\" in history text\n", p);
+	    } else {
+	        fprintf(stderr, "Can't store \"%s\", %s\n",
+		        p, strerror(errno));
+	        if (temp[0])
+		    (void)unlink(temp);
+	        exit(1);
+	    }
 	}
     }
     if (QIOerror(qp)) {
